Split Session definitions to source file
